Aggregate Pattern #2: Separate wrapper Class for methods in Itegrator class of V2
Creating a 2 wrapper class get_data_v1 and get_data_v2 in det_data.py file, to wrap all the methods for API version 1 and 2 into two different class. The forms a aggregate pattern as the files that imports methods of get_data.py will now import classes of that file and will have to call the class in order to use a particular method. hence class forms a has-a relationship with its method also separating version 1 and 2 into different groups for better understanding