Making DataProvider class for get_data file
It is correct that many data providers for this app is John Hopkins University but, we should create the app such that it is able to get data from different sources in case that later there are more sources of data for covid cases. To do this, we should make a data provider class and receives that data source name and URLs and applies different get data functions based on these inputs.