Adapter Pattern for Different API Versions
The API can have different versions. Users might be interested in calling functions from different versions. Thus, we add a structural pattern to adapt different versions of the API. Our pattern is an Adapter pattern which is implemented via a class that has an object for each version of the API and the right one will be called based on the user's request. Some of these function are added to the APIversionAdapter class.