Related Issue:
    N/A

Proposed Changes:

My proposed change is to make an adapter that allows for writing logs to UDP sockets instead of only having the ability to write to files. This gives more options on how to save logs.

Additional Info

UDP sockets are somethign I am just starting to learn about, so the code is based around them. I believe the same idea and methedology could be applied if writing to something other then UDP sockets wanted to be implimented.

Checklist

    [x]Tests
    [x]Translations
    [x]Documentation